when deleting, temp replace of targetrange by model-selection and shiftedVisually
