+++ addon/plugins/list/list.ts
@@ -1,4 +1,7 @@
-import { eventTargetRange } from '@lblod/ember-rdfa-editor/input/utils';
+import {
+  deleteTargetRange,
+  eventTargetRange,
+} from '@lblod/ember-rdfa-editor/input/utils';
 import Controller from '@lblod/ember-rdfa-editor/model/controller';
 import ModelElement from '@lblod/ember-rdfa-editor/model/model-element';
 import ModelNode from '@lblod/ember-rdfa-editor/model/model-node';
@@ -31,11 +34,7 @@ export default class ListPlugin implements EditorPlugin {
   }
 
   handleDelete(event: InputEvent, direction: number) {
-    const range = eventTargetRange(
-      this.controller.currentState,
-      this.controller.view.domRoot,
-      event
-    );
+    const range = deleteTargetRange(this.controller.currentState, direction);
     if (range.collapsed && direction === -1) {
       const lis = [
         ...range.end.parent.findSelfOrAncestors(ModelNodeUtils.isListElement),
